requirements.txt
pandas
numpy
setuptools
google
google-cloud
google-cloud-storage
PyYAML